One Vs All Regularized Logistic Regression
Simple Python implementation of the One Vs All Regularized Logistic Regression multi class classifier for the digit recognition problem. 

Do To: build models in parallel on multi-core systems and Test it against the data set provided on Kaggle.  
None of the tuning parameters (in particular lambda and the number of iterations for Gradient Descent) has been optimized for better performance 

The CSV is being used for testing the algorithm while developing.